+++ NEWS.md
@@ -1,7 +1,23 @@
 # Version 0.0 - Release Notes
 
+* [0.2.6] 
+    * released 2023-12-01
+    * sp_build_rsp_x update; added further value handling 
+    * summary.respeciate update; removed output print
+    * pls_rebuild update; option to add species as marker  
+    * now importing lattice and latticeExtra for plot panelling
+    * pls_plots and plot.respeciate now using lattice/latticeExtra 
+    * added sp_plot_profile and sp_profile_species 
+    * reduced padding on log scale y-axis  
+
+* [0.2.5] 
+    * released 2023-10-25
+    * sp_build_rsp_x update; x converted to data.frame (trouble with tibbles!)
+    * pls_plot update; changed mar settings 
+    * added first draft pls_test 
+
 * [0.2.4] 
-    * released 2023-11-14 
+    * released 2023-10-14 
     * added functions: pls_fit_species 
     * added draft rsp_pls and rsp_x object classes 
     * added draft sp_build_rsp_x